added image crop for editor photos
