+++ src/client/components/EditorExtended/components/sides/ImageSideButton.js
@@ -79,6 +79,7 @@ class ImageSideButton extends React.Component {
           onOk={this.handleOnOk}
         >
           <ImageSetter
+            isEditable
             onImageLoaded={this.getImages}
             onLoadingImage={this.onLoadingImage}
             isRequired
