+++ src/client/components/ImageSetter/ImageSetter.js
@@ -1,12 +1,10 @@
 import React, { useEffect, useRef, useState } from 'react';
 import PropTypes from 'prop-types';
-import { FormattedMessage, injectIntl } from 'react-intl';
-import { Icon, message, Slider } from 'antd';
-import AvatarEditor from 'react-avatar-editor';
+import { injectIntl } from 'react-intl';
+import { Icon, message } from 'antd';
 import { map, isEmpty, get, isEqual, isNil, size } from 'lodash';
 import { EditorState } from 'draft-js';
 import uuidv4 from 'uuid/v4';
-import fetch from 'isomorphic-fetch';
 import classNames from 'classnames';
 import withEditor from '../Editor/withEditor';
 import { isValidImage } from '../../../common/helpers/image';
@@ -17,9 +15,9 @@ import {
 } from '../../../common/constants/validation';
 import { objectFields } from '../../../common/constants/listOfFields';
 import { hexToRgb } from '../../../common/helpers';
-
-import './ImageSetter.less';
 import useWebsiteColor from '../../../hooks/useWebsiteColor';
+import ImageSetterEditor from './ImageSetterEditor';
+import './ImageSetter.less';
 
 const ImageSetter = ({
   intl,
@@ -49,7 +47,6 @@ const ImageSetter = ({
   const [currentImages, setCurrentImages] = useState([]);
   const [isLoadingImage, setLoadingImage] = useState(false);
   const [fileImages, setFileImages] = useState([]);
-  const [editor, setEditor] = useState('');
   const [isOpen, setIsOpen] = useState(false);
   const initialState = {
     image: '',
@@ -67,43 +64,9 @@ const ImageSetter = ({
 
   const colors = useWebsiteColor();
 
-  const handlePositionChange = position => {
-    setState({ ...state, position });
-  };
-
   const handleNewImage = async e => {
     await setState({ ...initialState, image: e.target.files[0] });
     setIsOpen(true);
-  };
-
-  const handleScale = scale => {
-    setState({ ...state, scale });
-  };
-
-  const rotateLeft = e => {
-    e.preventDefault();
-    setState({ ...state, rotate: state.rotate - 90 });
-  };
-  const rotateRight = e => {
-    e.preventDefault();
-    setState({ ...state, rotate: state.rotate + 90 });
-  };
-  const setEditorRef = ed => {
-    setEditor(ed);
-  };
-  const handleSave = async () => {
-    const dataUrl = editor.getImage().toDataURL();
-    const result = await fetch(dataUrl);
-    const blob = await result.blob();
-
-    const res = new File([blob], 'filename', { type: 'image/png' });
-
-    handleChangeImage({ target: { files: [res] } });
-    setIsOpen(false);
-  };
-  const resetImage = () => {
-    setState(initialState);
-    setIsOpen(false);
   };
 
   useEffect(() => {
@@ -442,53 +405,14 @@ const ImageSetter = ({
           </div>
         </div>
       )}
-      {isOpen && (
-        <div className="ImageSetter__edit">
-          <div className="image-box__preview">
-            <div className="image-box__remove" onClick={() => resetImage()} role="presentation">
-              <i className="iconfont icon-delete_fill Image-box__remove-icon" />
-            </div>
-            <AvatarEditor
-              ref={setEditorRef}
-              scale={parseFloat(state.scale)}
-              width={state.width}
-              height={state.height}
-              position={state.position}
-              onPositionChange={handlePositionChange}
-              rotate={parseFloat(state.rotate)}
-              borderRadius={state.width / (100 / state.borderRadius)}
-              image={state.image}
-            />
-          </div>
-          <div className="ImageSetter__zoom">
-            <div>
-              <FormattedMessage id="zoom" defaultMessage="Zoom" />:
-            </div>
-            <Slider
-              tipFormatter={null}
-              defaultValue={1}
-              min={state.allowZoomOut ? 0.1 : 1}
-              max={3}
-              step={0.01}
-              onChange={handleScale}
-            />
-          </div>
-          <div className="ImageSetter__rotate">
-            <FormattedMessage id="rotate" defaultMessage="Rotate" />:
-            <div className="ImageSetter__rotate-btns">
-              <button className="ImageSetter__rotate__button" onClick={rotateLeft}>
-                <FormattedMessage id="left" defaultMessage="Left" />
-              </button>
-              <button className="ImageSetter__rotate__button" onClick={rotateRight}>
-                <FormattedMessage id="right" defaultMessage="Right" />
-              </button>
-            </div>
-          </div>
-          <button onClick={handleSave} type="button" className="ImageSetter__save-btn">
-            <FormattedMessage id="save" defaultMessage="Save" />
-          </button>
-        </div>
-      )}
+      <ImageSetterEditor
+        state={state}
+        setState={setState}
+        isOpen={isOpen}
+        setIsOpen={setIsOpen}
+        initialState={initialState}
+        handleChangeImage={handleChangeImage}
+      />
     </div>
   );
 };
